Align type used to represent decimal digits (#46)

+++ include/roq/market/market_by_price.h
@@ -24,8 +24,8 @@ class ROQ_PUBLIC MarketByPrice {
   virtual double quantity_increment() const = 0;
 
   // precision required to show all significant decimal places (-1 means undefined)
-  virtual int32_t price_decimal_digits() const = 0;
-  virtual int32_t quantity_decimal_digits() const = 0;
+  virtual int8_t price_decimal_digits() const = 0;
+  virtual int8_t quantity_decimal_digits() const = 0;
 
   // convert price/quantity to internal representation
   // note! some events can change the internal representation, e.g. ReferenceData
